requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.0
PyJWT==2.8.0
Werkzeug==2.3.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
numpy==1.24.0
pandas==2.0.0
scikit-learn==1.2.0
